Rename functionaltests -> integrationservertest
