+++ .github/workflows/integration-server-test.yml
@@ -1,5 +1,5 @@
 ---
-name: functional-tests
+name: integration-server-test
 
 on:
   workflow_dispatch: ~
@@ -45,7 +45,7 @@ jobs:
           terraform_version: "${{ env.TERRAFORM_VERSION }}"
       - uses: actions/setup-go@0aaccfd150d50ccaeb58ebd88d36e91967a5f35b # v5
         with:
-          go-version-file: 'functionaltests/go.mod'
+          go-version-file: 'integrationservertest/go.mod'
           cache: false
       - uses: elastic/oblt-actions/google/auth@db0844c27572a45966ce669e4d3073c205c8d5e4 # v1
       - uses: google-github-actions/get-secretmanager-secrets@a8440875e1c2892062aef9061228d4f1af8f919b # v2.2.3
@@ -54,7 +54,7 @@ jobs:
           secrets: |-
             EC_API_KEY:elastic-observability/elastic-cloud-observability-team-${{ matrix.environment }}-api-key
       - name: "Run upgrade tests"
-        working-directory: ${{ github.workspace }}/functionaltests
+        working-directory: ${{ github.workspace }}/integrationservertest
         run: |
           export TF_VAR_CREATED_DATE=$(date +%s)
           go test -run="TestUpgrade.*/.*/${{ matrix.scenario }}" -v -timeout=60m -target="${{ matrix.environment }}" -upgrade-path="${{ matrix.upgrade-path }}" ./
@@ -78,7 +78,7 @@ jobs:
           terraform_version: "${{ env.TERRAFORM_VERSION }}"
       - uses: actions/setup-go@0aaccfd150d50ccaeb58ebd88d36e91967a5f35b # v5
         with:
-          go-version-file: 'functionaltests/go.mod'
+          go-version-file: 'integrationservertest/go.mod'
           cache: false
       - uses: elastic/oblt-actions/google/auth@db0844c27572a45966ce669e4d3073c205c8d5e4 # v1
       - uses: google-github-actions/get-secretmanager-secrets@a8440875e1c2892062aef9061228d4f1af8f919b # v2.2.3
@@ -87,7 +87,7 @@ jobs:
           secrets: |-
             EC_API_KEY:elastic-observability/elastic-cloud-observability-team-${{ matrix.environment }}-api-key
       - name: "Run standalone-to-managed tests"
-        working-directory: ${{ github.workspace }}/functionaltests
+        working-directory: ${{ github.workspace }}/integrationservertest
         run: |
           export TF_VAR_CREATED_DATE=$(date +%s)
           go test -run="TestStandaloneManaged.*/${{ matrix.scenario }}" -v -timeout=60m -target="${{ matrix.environment }}" ./
